requirements.txt
Flask
Flask-CORS
requests
python-dotenv
langchain 
openai
langchain_community
langchain_groq
